Refactor reader code
- Move all rasterio specific code into
  `odc.stac._rio` module
